move user interaction code from Program.cs to AskUser static class
